src/Twitch Tools/requirements.txt
requests==2.26.0
websocket-client==1.2.1
regex==2021.8.28
emoji==1.4.2
twitchio==2.0.6